stackanalysis, stack alignment, overwrite loops, intra-syslib calls
- Restores liveness for non-malware binaries
- Re-enables stack-alignment so that any calls made by instrumentation
will occur on a word-aligned stack (Ray's stuff)
- Doesn't call parseAfterCallAndInstrument for intramodule calls
from syslib to syslib
- Fixes bug in multi-function overwrite loop selection, ensures
that the selected overwrite loop won't trigger further analysis
of its own code, and cleans up dead code
- Changes cerrs to malware_cerrs

+++ dyninstAPI/src/BPatch_asyncEventHandler.C
@@ -643,7 +643,7 @@ bool BPatch_asyncEventHandler::waitNextEvent(EventRecord &ev)
      event_queue.push_back(ev);
   }
 
-  int width = 0;
+  unsigned width = 0;
   fd_set readSet;
   fd_set errSet;
 
@@ -658,7 +658,7 @@ bool BPatch_asyncEventHandler::waitNextEvent(EventRecord &ev)
 #if defined (os_windows)
   FD_SET(windows_sock, &readSet);
   FD_SET(windows_sock, &errSet);
-  if (windows_sock > (unsigned)width)
+  if (windows_sock > width)
       width = windows_sock;
 #else
   //  build the set of fds we want to wait on, one fd per process
